+++ src/main/java/com/team/comma/domain/track/track/controller/TrackController.java
@@ -1,7 +1,5 @@
 package com.team.comma.domain.track.track.controller;
 
-import com.team.comma.domain.track.track.dto.TrackRequest;
-import com.team.comma.domain.track.track.service.TrackService;
 import com.team.comma.domain.track.track.service.TrackService;
 import com.team.comma.global.common.dto.MessageResponse;
 import lombok.RequiredArgsConstructor;
@@ -15,21 +13,6 @@ import javax.security.auth.login.AccountException;
 public class TrackController {
 
     private final TrackService trackService;
-
-    @PatchMapping("/play-count/{trackId}")
-    public MessageResponse countPlayCount(@CookieValue String accessToken , @PathVariable String trackId) throws AccountException {
-        return trackService.countPlayCount(accessToken , trackId);
-    }
-
-    @GetMapping
-    public MessageResponse findMostListenedTrack(@CookieValue String accessToken) {
-        return trackService.findMostListenedTrack(accessToken);
-    }
-
-    @GetMapping("/friends")
-    public MessageResponse findMostListenedTrackByFriend(@CookieValue String accessToken) {
-        return trackService.findMostListenedTrackByFriend(accessToken);
-    }
 
     @GetMapping("/users/favorites")
     public MessageResponse findTrackByFavoriteTrack(@CookieValue String accessToken) {
