+++ src/test/java/com/team/comma/domain/track/track/controller/TrackControllerTest.java
@@ -73,176 +73,6 @@ public class TrackControllerTest {
     }
 
     @Test
-    @DisplayName("트랙 추천 수 증가 실패 _ 트랙 탐색 실패")
-    public void countPlayCountFail_notFountTrack() throws Exception {
-        // given
-        final String url = "/tracks/play-count/{trackId}";
-        doThrow(new TrackException("트랙을 찾을 수 없습니다.")).when(trackService).countPlayCount(any(String.class) , any(String.class));
-
-        // when
-        final ResultActions resultActions = mockMvc.perform(
-                RestDocumentationRequestBuilders.patch(url , "trackId")
-                        .cookie(new Cookie("accessToken", "accessToken"))
-        );
-
-        // then
-        resultActions.andExpect(status().isBadRequest()).andDo(
-                document("track/addTrackListenedCountFail-notFountTrack",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        pathParameters(
-                                parameterWithName("trackId").description("스포티파이 트랙 ID")
-                        ),
-                        requestCookies(
-                                cookieWithName("accessToken").description("사용자 access token")
-                        ),
-                        responseFields(
-                                fieldWithPath("code").description("응답 코드"),
-                                fieldWithPath("message").description("메세지"),
-                                fieldWithPath("data").description("데이터")
-                        )
-                )
-        );
-        final MessageResponse result = gson.fromJson(
-                resultActions.andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8),
-                MessageResponse.class);
-
-        assertThat(result.getCode()).isEqualTo(SIMPLE_REQUEST_FAILURE.getCode());
-        assertThat(result.getMessage()).isEqualTo("트랙을 찾을 수 없습니다.");
-    }
-
-    @Test
-    @DisplayName("트랙 추천 수 증가")
-    public void countPlayCount() throws Exception {
-        // given
-        final String url = "/tracks/play-count/{trackId}";
-        doReturn(MessageResponse.of(REQUEST_SUCCESS)).when(trackService).countPlayCount(any(String.class) , any(String.class));
-
-        // when
-        final ResultActions resultActions = mockMvc.perform(
-                RestDocumentationRequestBuilders.patch(url , "trackId")
-                        .cookie(new Cookie("accessToken", "accessToken"))
-        );
-
-        // then
-        resultActions.andExpect(status().isOk()).andDo(
-                document("track/addTrackListenedCount",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        pathParameters(
-                                parameterWithName("trackId").description("스포티파이 트랙 ID")
-                        ),
-                        requestCookies(
-                                cookieWithName("accessToken").description("사용자 access token")
-                        ),
-                        responseFields(
-                                fieldWithPath("code").description("응답 코드"),
-                                fieldWithPath("message").description("메세지"),
-                                fieldWithPath("data").description("데이터")
-                        )
-                )
-        );
-        final MessageResponse result = gson.fromJson(
-                resultActions.andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8),
-                MessageResponse.class);
-
-        assertThat(result.getCode()).isEqualTo(REQUEST_SUCCESS.getCode());
-        assertThat(result.getMessage()).isEqualTo(REQUEST_SUCCESS.getMessage());
-    }
-
-    @Test
-    @DisplayName("내가 가장 많이 들은 곡")
-    public void findMostListenedTrack() throws Exception {
-        // given
-        final String url = "/tracks";
-        List<TrackPlayCountResponse> trackPlayCountResponses = new ArrayList<>();
-        for(int i = 0; i < 2; i++) {
-            trackPlayCountResponses.add(buildTrackPlayCountResponse());
-        }
-
-        doReturn(MessageResponse.of(REQUEST_SUCCESS , trackPlayCountResponses)).when(trackService).findMostListenedTrack(any(String.class));
-
-        // when
-        final ResultActions resultActions = mockMvc.perform(
-                MockMvcRequestBuilders.get(url)
-                        .cookie(new Cookie("accessToken", "accessToken"))
-        );
-
-        // then
-        resultActions.andExpect(status().isOk()).andDo(
-                document("track/mostListenTrackByMe",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        requestCookies(
-                                cookieWithName("accessToken").description("사용자 access token")
-                        ),
-                        responseFields(
-                                fieldWithPath("code").description("응답 코드"),
-                                fieldWithPath("message").description("메세지"),
-                                fieldWithPath("data").description("데이터"),
-                                fieldWithPath("data.[].playCount").description("재생 횟수"),
-                                fieldWithPath("data.[].trackId").description("스포티파이 트랙 ID"),
-                                fieldWithPath("data.[].trackImageUrl").description("트랙 이미지 URL"),
-                                fieldWithPath("data.[].trackName").description("트랙 이름"),
-                                fieldWithPath("data.[].trackArtist").description("트랙 아티스트")
-                        )
-                )
-        );
-        final MessageResponse result = gson.fromJson(
-                resultActions.andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8),
-                MessageResponse.class);
-
-        assertThat(result.getCode()).isEqualTo(REQUEST_SUCCESS.getCode());
-        assertThat(result.getMessage()).isEqualTo(REQUEST_SUCCESS.getMessage());
-    }
-
-    @Test
-    @DisplayName("친구가 가장 많이 들은 곡")
-    public void findMostListenedTrackByFriend() throws Exception {
-        // given
-        final String url = "/tracks/friends";
-        List<TrackPlayCountResponse> trackPlayCountResponses = new ArrayList<>();
-        for(int i = 0; i < 2; i++) {
-            trackPlayCountResponses.add(buildTrackPlayCountResponse());
-        }
-
-        doReturn(MessageResponse.of(REQUEST_SUCCESS , trackPlayCountResponses)).when(trackService).findMostListenedTrackByFriend(any(String.class));
-
-        // when
-        final ResultActions resultActions = mockMvc.perform(
-                MockMvcRequestBuilders.get(url)
-                        .cookie(new Cookie("accessToken", "accessToken"))
-        );
-
-        // then
-        resultActions.andExpect(status().isOk()).andDo(
-                document("track/mostListenTrackByFriend",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        requestCookies(
-                                cookieWithName("accessToken").description("사용자 access token")
-                        ),
-                        responseFields(
-                                fieldWithPath("code").description("응답 코드"),
-                                fieldWithPath("message").description("메세지"),
-                                fieldWithPath("data").description("데이터"),
-                                fieldWithPath("data.[].playCount").description("재생 횟수"),
-                                fieldWithPath("data.[].trackId").description("스포티파이 트랙 ID"),
-                                fieldWithPath("data.[].trackImageUrl").description("트랙 이미지 URL"),
-                                fieldWithPath("data.[].trackName").description("트랙 이름"),
-                                fieldWithPath("data.[].trackArtist").description("트랙 아티스트")
-                        )
-                )
-        );
-        final MessageResponse result = gson.fromJson(
-                resultActions.andReturn().getResponse().getContentAsString(StandardCharsets.UTF_8),
-                MessageResponse.class);
-
-        assertThat(result.getCode()).isEqualTo(REQUEST_SUCCESS.getCode());
-        assertThat(result.getMessage()).isEqualTo(REQUEST_SUCCESS.getMessage());
-    }
-
-    @Test
     @DisplayName("유저가 좋아요 누른 트랙 탐색")
     public void findTrackByFavoriteTrack() throws Exception {
         // given
@@ -342,15 +172,7 @@ public class TrackControllerTest {
         assertThat(result.getMessage()).isEqualTo(REQUEST_SUCCESS.getMessage());
     }
 
-    public TrackPlayCountResponse buildTrackPlayCountResponse() {
-        return TrackPlayCountResponse.builder()
-                .playCount(0)
-                .trackId("trackId")
-                .trackImageUrl("images")
-                .trackName("trackName")
-                .trackArtist("trackArtist")
-                .build();
-    }
+
 
     private Track buildTrack(String title, String spotifyId) {
         return Track.builder()
